requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.1
boto3==1.35.97
google-cloud-storage==2.19.0